refactor : 엔드포인트 규칙에 맞게 수정, 그에 따른 테스트 수정
