+++ src/main/java/com/hansarangdelivery/controller/CategoryController.java
@@ -44,7 +44,7 @@ public class CategoryController {
         return new ResultResponseDto<>("카테고리 삭제 성공",200, result);
     }
 
-    @GetMapping("/search")
+    @GetMapping("/")
     public ResultResponseDto<Page<CategoryResponseDto>> searchCategories(
         Pageable pageable
         ) {
