+++ src/main/java/com/hansarangdelivery/controller/RestaurantController.java
@@ -65,9 +65,9 @@ public class RestaurantController {
         return new ResultResponseDto<>("가게 삭제 성공", 200,result);
     }
 
-    @GetMapping("/search")
+    @GetMapping("/")
     public ResultResponseDto<Page<RestaurantResponseDto>> searchRestaurants(
-    // 레스토랑 리스트 검색, search가 포함된 음식점을 찾는데 사용될 예정
+    // 레스토랑 리스트 검색
         Pageable pageable,
         @RequestParam(required = false) String search,
         @RequestParam(required = false) String category) {
